# LAB6/static/swagger.json
{
    "openapi": "3.0.0",

    "info": {
        "title": "SCOOTER API",
        "version": "1.0.0"
    },

    "tags": [
        {
          "name": "Electro-scooter API"
        }
    ],

    "components": {
        "schemas": {
            "Scooter": {
                "type": "object",
                "properties": {
                    "name": { "type": "string" },
                    "battery_level": { "type": "number" }
                }
            },
            "Id":{
                "type" : "integer",
                "properties" : {
                    "name" : { "type" : "integer" }
                }
            }
        }
    },

    "paths": {
        "/api/electro-scooters": {
            "post": {
                "tags": [ "Add a new scooter"],
                "summary": "Add a new scooter",
                "requestBody": {
                    "description": "New scooter",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Scooter"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "schema": {
                            "$ref": "#/components/schemas/Scooter"
                        }
                    },
                    "400": {
                        "description": "BAD REQUEST",
                        "schema": {
                            "$ref": "#/components/schemas/Scooter"
                        }
                    }
                }
            }
        },

        "/api/electro-scooters/{id}": {
            "get" :{
                "tags": [ "Get the data of a scooter" ],
                "summary": "Get the data of a scooter",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "Scooter id",
                    "required": true,
                    "type": "integer",
                    "schema": {
                        "$ref": "#/components/schemas/Id"
                    }
                }],

                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Scooter"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Scooter"
                        }
                    }
                }
            },

            "put": {
                "tags": [ "Update the data of a scooter" ],
                "summary": "Update the data of a scooter",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "Scooter id",
                    "required": true,
                    "type": "integer",
                    "schema": {
                        "$ref": "#/components/schemas/Id"
                    }
                }],

                "requestBody": {
                    "description": "Update the data of a scooter",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Scooter"
                            }
                        }
                    }
                },

                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Scooter"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Scooter"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR",
                        "schema": {
                            "$ref": "#/components/schemas/Scooter"
                        }
                    }
                }
            },

            "delete": {
                "tags": [ "Delete a scooter" ],
                "summary": "Delete a scooter",

                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Scooter id",
                        "required": true,
                        "type": "integer",
                        "schema": {
                            "$ref": "#/components/schemas/Id"
                        }
                    },
                    {
                        "name": "X-Delete-Password",
                        "in": "header",
                        "description": "Password",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],

                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/components/schemas/Scooter"
                        }
                    },
                    "401": {
                        "description": "UNAUTHORIZED",
                        "schema":{
                            "$ref": "#/components/schemas/Scooter"
                        }
                    },
                    "404": {
                        "description": "NOT FOUND",
                        "schema": {
                            "$ref": "#/components/schemas/Scooter"
                        }
                    },
                    "500": {
                        "description": "INTERNAL SERVER ERROR",
                        "schema": {
                            "$ref": "#/components/schemas/Scooter"
                        }
                    }
                }
            }
        }
    }
}